refactor: move `units` to `cssHelper`
