+++ apps/calendar/src/components/timeGrid/gridSelection.tsx
@@ -2,8 +2,7 @@ import { FunctionComponent, h } from 'preact';
 
 import { GridSelectionInfo, timeFormats } from '@src/components/timeGrid';
 import { toFormat } from '@src/time/datetime';
-import { cls } from '@src/util/cssHelper';
-import { toPercent } from '@src/util/units';
+import { cls, toPercent } from '@src/util/cssHelper';
 
 interface Props extends GridSelectionInfo {
   top: number;
